renamed gate_constrcution
